# apps/frontend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: StipendieAssistenten Backend API
  description: |
    Required backend endpoints for StipendieAssistenten frontend.
    All authenticated endpoints require a Bearer token in the Authorization header.
  version: 1.0.0
servers:
  - url: /api
    description: Default API path (configure via VITE_API_URL env variable)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FamilyMember:
      type: object
      required:
        - id
        - name
        - age
        - role
      properties:
        id:
          type: string
        name:
          type: string
        age:
          type: integer
        role:
          type: string
          enum: [adult, child]
    ChildNeed:
      type: object
      properties:
        childId:
          type: string
        diagnoses:
          type: array
          items:
            type: string
            enum: [adhd, autism, cp, mobility_impairment, other]
        needDegree:
          type: integer
          minimum: 0
          maximum: 3
        otherDiagnosis:
          type: string
          nullable: true
    FamilyProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        municipality:
          type: string
        maritalStatus:
          type: string
          enum: [single, married, cohabiting, divorced, widowed]
        familyMembers:
          type: array
          items:
            $ref: '#/components/schemas/FamilyMember'
        childNeeds:
          type: array
          items:
            $ref: '#/components/schemas/ChildNeed'
        helpCategories:
          type: array
          items:
            type: string
            enum: [activities, home_adaptation, travel, aids, economy, education, special_projects]
        receivedSupport:
          type: object
          properties:
            careAllowance:
              type: boolean
            lssSupport:
              type: boolean
            shortTermStay:
              type: boolean
            housingBenefits:
              type: boolean
        economicSituation:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /profile/family:
    get:
      summary: Get user's family profile
      tags: [Profile]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Family profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyProfile'
        '404':
          description: No profile found (user should create one)
        '401':
          description: Not authenticated

    put:
      summary: Create or update family profile
      tags: [Profile]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyProfile'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyProfile'
        '400':
          description: Invalid profile data
        '401':
          description: Not authenticated

  # ============================================
  # GRANTS (Public + Authenticated features)
  # ============================================
...
  /profile/saved-grants:
    get:
      summary: Get user's saved/favorite grants
      tags: [Profile, Grants]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of saved grant IDs
          content:
            application/json:
              schema:
                type: object
                properties:
                  saved_grants:
                    type: array
                    items:
                      type: string
        '401':
          description: Not authenticated

    post:
      summary: Save a grant to favorites
      tags: [Profile, Grants]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - grant_id
              properties:
                grant_id:
                  type: string
      responses:
        '201':
          description: Grant saved successfully
        '401':
          description: Not authenticated
  /profile/saved-grants/{grantId}:
    delete:
      summary: Remove a grant from favorites
      tags: [Profile, Grants]
      security:
        - BearerAuth: []
      parameters:
        - name: grantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Grant removed from favorites
        '401':
          description: Not authenticated

  # ============================================
  # APPLICATIONS (Authenticated)
  # ============================================
...